add vecPoly2 and PolyDeg3 ops to util file
