+++ src/range_proof.rs
@@ -17,13 +17,10 @@ use inner_product_proof;
 
 use proof_transcript::ProofTranscript;
 
-use util;
+use util::{self, PolyDeg3, VecPoly2};
 
 use generators::{Generators, GeneratorsView};
 
-struct PolyDeg3(Scalar, Scalar, Scalar);
-
-struct VecPoly2(Vec<Scalar>, Vec<Scalar>);
 
 /// The `RangeProof` struct represents a single range proof.
 #[derive(Clone, Debug)]
@@ -274,61 +271,6 @@ fn delta(n: usize, y: &Scalar, z: &Scalar) -> Scalar {
     (z - zz) * sum_of_powers_of_y - z * zz * sum_of_powers_of_2
 }
 
-impl VecPoly2 {
-    pub fn zero(n: usize) -> VecPoly2 {
-        VecPoly2(vec![Scalar::zero(); n], vec![Scalar::zero(); n])
-    }
-
-    pub fn inner_product(&self, rhs: &VecPoly2) -> PolyDeg3 {
-        // Uses Karatsuba's method
-        let l = self;
-        let r = rhs;
-
-        let t0 = inner_product(&l.0, &r.0);
-        let t2 = inner_product(&l.1, &r.1);
-
-        let l0_plus_l1 = add_vec(&l.0, &l.1);
-        let r0_plus_r1 = add_vec(&r.0, &r.1);
-
-        let t1 = inner_product(&l0_plus_l1, &r0_plus_r1) - t0 - t2;
-
-        PolyDeg3(t0, t1, t2)
-    }
-
-    pub fn eval(&self, x: Scalar) -> Vec<Scalar> {
-        let n = self.0.len();
-        let mut out = vec![Scalar::zero(); n];
-        for i in 0..n {
-            out[i] += self.0[i] + self.1[i] * x;
-        }
-        out
-    }
-}
-
-pub fn inner_product(a: &[Scalar], b: &[Scalar]) -> Scalar {
-    let mut out = Scalar::zero();
-    if a.len() != b.len() {
-        // throw some error
-        println!("lengths of vectors don't match for inner product multiplication");
-    }
-    for i in 0..a.len() {
-        out += a[i] * b[i];
-    }
-    out
-}
-
-pub fn add_vec(a: &[Scalar], b: &[Scalar]) -> Vec<Scalar> {
-    let mut out = Vec::new();
-    if a.len() != b.len() {
-        // throw some error
-        println!("lengths of vectors don't match for vector addition");
-    }
-    for i in 0..a.len() {
-        out.push(a[i] + b[i]);
-    }
-    out
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -358,23 +300,6 @@ mod tests {
         }
 
         assert_eq!(power_g, delta(n, &y, &z),);
-    }
-
-    #[test]
-    fn test_inner_product() {
-        let a = vec![
-            Scalar::from_u64(1),
-            Scalar::from_u64(2),
-            Scalar::from_u64(3),
-            Scalar::from_u64(4),
-        ];
-        let b = vec![
-            Scalar::from_u64(2),
-            Scalar::from_u64(3),
-            Scalar::from_u64(4),
-            Scalar::from_u64(5),
-        ];
-        assert_eq!(Scalar::from_u64(40), inner_product(&a, &b));
     }
 
     fn create_and_verify_helper(n: usize) {
